+++ spec/integration/sim_control_spec.rb
@@ -4,64 +4,6 @@ describe RunLoop::SimControl do
 
   subject(:sim_control) { RunLoop::SimControl.new }
 
-  # flickering on Travis CI
-  unless Resources.shared.travis_ci?
-    describe '#quit_sim and #launch_sim' do
-      before(:each) { RunLoop::SimControl.terminate_all_sims }
-
-      it "with Xcode #{Resources.shared.current_xcode_version}" do
-        sim_control.launch_sim
-        expect(sim_control.sim_is_running?).to be == true
-
-        sim_control.quit_sim
-        expect(sim_control.sim_is_running?).to be == false
-      end
-
-      xcode_installs = Resources.shared.alt_xcode_install_paths
-      unless xcode_installs.empty?
-        describe 'regression' do
-          xcode_installs.each do |developer_dir|
-            it "#{developer_dir}" do
-              Resources.shared.with_developer_dir(developer_dir) do
-                local_sim_control = RunLoop::SimControl.new
-                local_sim_control.launch_sim
-                expect(local_sim_control.sim_is_running?).to be == true
-
-                local_sim_control.quit_sim
-                expect(local_sim_control.sim_is_running?).to be == false
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-
-  describe '#relaunch_sim' do
-
-    before(:each) { RunLoop::SimControl.terminate_all_sims }
-
-    it "with Xcode #{Resources.shared.current_xcode_version}" do
-      sim_control.relaunch_sim
-      expect(sim_control.sim_is_running?).to be == true
-    end
-
-    xcode_installs = Resources.shared.alt_xcode_install_paths
-    unless xcode_installs.empty?
-      describe 'regression' do
-        xcode_installs.each do |developer_dir|
-          it "#{developer_dir}" do
-            Resources.shared.with_developer_dir(developer_dir) do
-              local_sim_control = RunLoop::SimControl.new
-              local_sim_control.relaunch_sim
-              expect(local_sim_control.sim_is_running?).to be == true
-            end
-          end
-        end
-      end
-    end
-  end
-
   describe '#sim_app_support_dir' do
     before(:each) {  RunLoop::SimControl.terminate_all_sims }
     it "with Xcode #{Resources.shared.current_xcode_version} returns a path that exists" do