+++ spec/lib/xcrun_spec.rb
@@ -32,12 +32,12 @@ describe RunLoop::Xcrun do
     end
 
     it "re-raises error if UTF8 encoding fails" do
-      error = RunLoop::Xcrun::UTF8Error.new("complex message")
-      expect(xcrun).to receive(:encode_utf8_or_raise).and_raise(error)
+      error = RunLoop::Encoding::UTF8Error.new("complex message")
+      expect(xcrun).to receive(:ensure_command_output_utf8).and_raise(error)
 
       expect do
         xcrun.exec(["sleep", "0.5"])
-      end.to raise_error RunLoop::Xcrun::UTF8Error, /complex message/
+      end.to raise_error RunLoop::Encoding::UTF8Error, /complex message/
     end
 
     it 're-raises error thrown by CommandRunner' do
@@ -65,81 +65,6 @@ describe RunLoop::Xcrun do
       end
     end
 
-    describe "#encode_utf8_or_raise" do
-      let(:string) { "string" }
-      let(:encoded) { "encoded" }
-      let(:forced) { "forced" }
-      let(:command) { "command" }
-
-      it "returns '' if string arg is falsey" do
-        expect(xcrun.send(:encode_utf8_or_raise, nil, command)).to be == ''
-      end
-
-      it "returns utf8 encoding" do
-        expect(string).to receive(:force_encoding).with("UTF-8").and_return(encoded)
-        expect(encoded).to receive(:chomp).and_return(encoded)
-        expect(encoded).to receive(:valid_encoding?).and_return(true)
-
-        expect(xcrun.send(:encode_utf8_or_raise, string, command)).to be == encoded
-      end
-
-      it "forces utf8 encoding" do
-        expect(string).to receive(:force_encoding).with("UTF-8").and_return(encoded)
-        expect(encoded).to receive(:chomp).and_return(encoded)
-        expect(encoded).to receive(:valid_encoding?).and_return(false)
-        expect(encoded).to receive(:encode).and_return(forced)
-        expect(forced).to receive(:valid_encoding?).and_return(true)
-
-        expect(xcrun.send(:encode_utf8_or_raise, string, command)).to be == forced
-      end
-
-      it "raises an error if string cannot be coerced to UTF8" do
-        expect(string).to receive(:force_encoding).with("UTF-8").and_return(encoded)
-        expect(encoded).to receive(:chomp).and_return(encoded)
-        expect(encoded).to receive(:valid_encoding?).and_return(false)
-        expect(encoded).to receive(:encode).and_return(forced)
-        expect(forced).to receive(:valid_encoding?).and_return(false)
-
-        expect do
-          xcrun.send(:encode_utf8_or_raise, string, command)
-        end.to raise_error RunLoop::Xcrun::UTF8Error,
-        /Could not force UTF-8 encoding on this string:/
-      end
-
-      it "handles string with non-UTF8 characters" do
-        file = "spec/resources/ps-with-non-utf8.log"
-        string = File.read(file)
-
-        version = RunLoop::Version.new(RUBY_VERSION)
-        if version < RunLoop::Version.new("2.1.0")
-           expect do
-             xcrun.send(:encode_utf8_or_raise, string, command)
-           end.to raise_error RunLoop::Xcrun::UTF8Error
-        else
-          actual = xcrun.send(:encode_utf8_or_raise, string, command)
-          split = actual.split($-0)
-
-          expect(split[0]).to be == "  PID COMMAND"
-          expect(split[1]).to be == "  324 /usr/libexec/UserEventAgent (Aqua)"
-          expect(split[2]).to be == "  403 /Applications/M^\\M^IM^AM^SM^MM^E.app/Contents/MacOS/M^\\M^IM^AM^SM^MM^E"
-          expect(split[3]).to be == " 1497 irb"
-        end
-      end
-
-      it "handles UTF-8 strings" do
-        # Force C (non UTF-8 encoding)
-        stub_env({'LC_ALL' => 'C'})
-        args = ['cat', 'spec/resources/encoding.txt']
-
-        # Confirm that the string is read as ASCII-US8BIT
-        command_runner_hash = CommandRunner.run(args, timeout: 0.2)
-        command_runner_out = command_runner_hash[:out]
-        expect(command_runner_out.length).to be == 22
-
-        actual = xcrun.send(:encode_utf8_or_raise, command_runner_out, command)
-        expect(actual).to be == 'ITZVÓÃ ●℆❡♡'
-      end
-    end
 
     describe 'contents of returned hash' do
       it 'mocked' do